xxx.size(): static_cast size_t to int32_t

+++ libstmm-input-gtk/src/flogtkbackend.cc
@@ -192,7 +192,7 @@ void GtkBackend::removeFloatingDevice(int32_t nIdx)
 {
 	onDeviceRemoved(m_aXDeviceIds[nIdx]);
 	//
-	const int32_t nLastIdx = m_aXDeviceIds.size() - 1;
+	const int32_t nLastIdx = static_cast<int32_t>(m_aXDeviceIds.size()) - 1;
 	if (nIdx < nLastIdx) {
 		m_aXDeviceIds[nIdx] = m_aXDeviceIds[nLastIdx];
 		m_aGdkDevices[nIdx] = m_aGdkDevices[nLastIdx];
